Add task conversion and promotion rules
Previously, trying to put different types of AbstractRealTimeTask into a
vector (or task system) would result in the collection having eltype
AbstractRealTimeTask.  This is inefficient, and not in line with the
normal philosophy of type promotion in Julia, since fundamentally the
different concrete task types store the same kind of thing.

To remedy this, I wrote convert and promote_rule methods to allow
automatic conversion and promotion of these types to each other.  Also
introduced are unit tests for these rules, in the
"Periodic[Implicit]Task promotion" test set.  PeriodicImplicitTasks are
promoted to PeriodicTasks, and their type parameters are promoted as one
would expect as well.  Very handy, and should encourage better use of
the type system.