Add support for sampling from MissRow constraints (#2)
* Add support for sampling from MissRow constraints

To support the statistical hypothesis testing work I've been involved in
recently, we need support for sampling from weakly hard constraints.
Uniform sampling is nicer to think about than a random walk, so we now
implement a fast algorithm for this (cited in the docs).

* Optimize MissRow sampling

Division with BigInts is slower than multiplication, so refactoring to
divide instead of multiplying speeds up the implementation even though
it causes more allocations.  Further, I realized there's no reason to
depend on Distributions, since we're really just checking if a random
value is less than a calculated probability.

* Rename SamplerMissRow→SamplerUniformMissRow

Better to specify that it's doing *uniform* sampling, since it's highly
likely we'll want to also implement random-walk sampling at some point.