Support iteration API over tasks and constraints
Task objects and weakly hard constraints now act as scalars for length
and iteration, so broadcasting across individual tasks and constraints
now works.  This allows us to compute things like demand_bound. and
request_bound. for a single task and a range of times.  We can also use
<=. broadcasting to compute things like Fig. 3 of Bernat, Burns, and
Llamosí in a single line.