Fix heuristic and WCRT analysis for JCL scheduling
There were two related problems here: one in the pseudocode in the
paper, and one in this implementation of it.  The paper neglected to
calculate the number of job-classes for hard real-time tasks as 1,
instead calculating it as window + 1.  This resulted in extra
job-classes, which in turn masked another bug in the WCRT analysis.  The
final condition "if p == K_k - m_k" is really checking if we're looking
at the last job-class, which may be index 1 if the job is hard
real-time.

Fixing these issues revealed a bug in my transcription of the
pseudocode, whereby I had forgotten that my p was off by 1 from the
paper's p.  This commit fixes both issues, making the analysis now (to
the best of my knowledge) fully correct.